Move the Context tests and benches into src/
This concludes preparations for making Context private again.
